+++ CHANGELOG.md
@@ -10,7 +10,8 @@ Versioning].
 
 ## [Unreleased]
 
-- _No changes yet_
+- ([#1565]) Allow template literals when `allowDerived` is set to true for the
+  `no-top-level-side-effects` rule.
 
 ## [3.5.3] - 2025-06-17
 
@@ -226,3 +227,4 @@ Versioning].
 [#1474]: https://github.com/ericcornelissen/eslint-plugin-top/pull/1474
 [#1475]: https://github.com/ericcornelissen/eslint-plugin-top/pull/1475
 [#1506]: https://github.com/ericcornelissen/eslint-plugin-top/pull/1506
+[#1565]: https://github.com/ericcornelissen/eslint-plugin-top/pull/1565