+++ CHANGELOG.md
@@ -12,6 +12,8 @@ Versioning].
 
 - ([#1565]) Allow template literals when `allowDerived` is set to true for the
   `no-top-level-side-effects` rule.
+- ([#1568]) Restrict unary operators to negative numbers when `allowDerived` for
+  the `no-top-level-side-effects` rule.
 
 ## [3.5.3] - 2025-06-17
 
@@ -228,3 +230,4 @@ Versioning].
 [#1475]: https://github.com/ericcornelissen/eslint-plugin-top/pull/1475
 [#1506]: https://github.com/ericcornelissen/eslint-plugin-top/pull/1506
 [#1565]: https://github.com/ericcornelissen/eslint-plugin-top/pull/1565
+[#1568]: https://github.com/ericcornelissen/eslint-plugin-top/pull/1568